requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
environ==1.0
fontawesome==5.10.1.post1
fontawesomefree==6.6.0
gunicorn==23.0.0
Markdown==3.7
mysqlclient==2.2.6
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
